Version 4
Tried to cut down the amount of numbers to loop through in findLargestPrimeFactor, so only looped up to the square root, but only checked to see if the input number/factor is prime and set that equal to the largestPrimeFactor.  However, this leaves out the possibility that the largestPrimeFactor is smaller than the square root.