Version 5/final
The largestPrimeFactor is more likely to be larger than the square root, so the number/i is checked to be prime first and sets largestPrimeFactor equal to number/i and breaks out of the loop.  Then if not, i is checked and, if prime, sets the value of largestPrimeFactor equal to i, but continues through the loop.  This ensures that there is no unnecessary looping and short circuits the function (if larger than the square root) when the right answer is found.  This function did not crash the browser and quickly returned the correct answer of 6857.